requirements.txt
pynsq
tornado